% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_macseq.R
\name{read_macseq}
\alias{read_macseq}
\title{Load in data from Mac-seq experiments}
\usage{
read_macseq(
  data_dir,
  label_file,
  gene_column = 2,
  cell_column = 1,
  unique_features = TRUE,
  strip_suffix = FALSE
)
}
\arguments{
\item{data_dir}{Directory containing the matrix.mtx, genes.tsv
(or features.tsv), and barcodes.tsv files provided by nextflow
pipeline using Starsolo.
Functionality for reading several data directories was removed
to allow reading in the metadata label matching to the Macseq
plate with minimum confusion.
Next version will consider integration of other modalities.}

\item{gene_column}{Specify which column of genes.tsv or features.tsv
to use for gene names; default is 2}

\item{cell_column}{Specify which column of barcodes.tsv
to use for cell names; default is 1}

\item{unique_features}{Make feature names unique (default TRUE)}

\item{strip_suffix}{Remove trailing "-1" if present in all cell barcodes.}

\item{label_dir}{Directory containing the matching metadata file in csv format.}
}
\value{
If features.csv indicates the data has multiple data types, a list
containing a sparse matrix of the data from each type will be returned.
Otherwise a sparse matrix containing the expression data will be returned.
}
\description{
This functions is a derivative from Read10X function from the Seurat package.
Enables easy loading of sparse data matrices provided by 10X genomics.
}
\examples{
\dontrun{
#
data_dir <- 'path/to/data/directory'
list.files(data_dir) # Should show barcodes.tsv, genes.tsv, and matrix.mtx
expression_matrix <- Readmacseq(data_dir = data_dir)
macpie_object = CreateMacPie(counts = expression_matrix)
}

}
\author{
Mark Li, \email{justdont}
}
\concept{preprocessing}
