% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMacSeq.R
\name{readMacSeq}
\alias{readMacSeq}
\title{Load in data from Mac-seq experiments}
\usage{
readMacSeq(
  data.dir,
  label.dir,
  gene.column = 2,
  cell.column = 1,
  unique.features = TRUE,
  strip.suffix = FALSE
)
}
\arguments{
\item{data.dir}{Directory containing the matrix.mtx, genes.tsv (or features.tsv), and barcodes.tsv
files provided by nextflow pipeline using Starsolo.
I removed the functionality for reading in several data directories as the original Read10X()
This is to allow me read in the metadata label matching to the Macseq plate with minimum confusion.
Next version will consider open opssibilities to integrate other omics data}

\item{label.dir}{Directory containing the matching metadata file in csv format.}

\item{gene.column}{Specify which column of genes.tsv or features.tsv to use for gene names; default is 2}

\item{cell.column}{Specify which column of barcodes.tsv to use for cell names; default is 1}

\item{unique.features}{Make feature names unique (default TRUE)}

\item{strip.suffix}{Remove trailing "-1" if present in all cell barcodes.}
}
\value{
If features.csv indicates the data has multiple data types, a list
containing a sparse matrix of the data from each type will be returned.
Otherwise a sparse matrix containing the expression data will be returned.
}
\description{
This functions is a derivative from Read10X() function from the Seurat package.
Enables easy loading of sparse data matrices provided by 10X genomics.
}
\examples{
\dontrun{
#
data_dir <- 'path/to/data/directory'
list.files(data_dir) # Should show barcodes.tsv, genes.tsv, and matrix.mtx
expression_matrix <- Readmacseq(data.dir = data_dir)
macpie_object = CreateMacPie(counts = expression_matrix)
}

}
\author{
Mark Li, \email{justdont}
}
\concept{preprocessing}
